Test only client.cube.slice in integration tests
Do only two simple integration tests:

1. Test that the RunTimeError 'Request timed out; unable to fetch
result' is raised when the client tries to fetch a slice from a cube
that does not exists. This error does not distinguish between a slice
that can not be fetched because the cube does not exists, or if a slice
can not be fetched because the blob with the results are not ready yet.
However, this is how things are for now.

2. Upload cube, fetch some slices and compare. For now, test only one
slice in each dimension. This will sometimes fail in the client with
RunTimeError 'Request timed out; unable to fetch result' because it
takes to much time to get the blob with the result ready.